requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.2
decorator==4.4.2
Django==4.0.2
django-4-jet==1.0.8
django-appconf==1.0.5
django-extra-views==0.14.0
gunicorn==20.1.0
idna==3.3
imageio==2.16.0
imageio-ffmpeg==0.4.5
moviepy==1.0.3
numpy==1.22.2
Pillow==9.0.0
proglog==0.1.9
pycparser==2.21
razorpay==1.3.0
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
tqdm==4.62.3
urllib3==1.26.8
